Organize redux (actions and reducers) by feature
