+++ features/actions/index.android.js
@@ -1,8 +1,3 @@
 require('../polyfills');
 
-export * from './client';
 export * from './tracks';
-export * from './participants';
-export * from './media';
-export * from './navigation';
-
