+++ features/constants.js
@@ -1,12 +1,4 @@
 export const JITSI_CLIENT_INIT = 'JITSI_CLIENT_INIT';
-export const JITSI_CLIENT_CREATED = 'JITSI_CLIENT_CREATED';
-export const JITSI_CLIENT_CONNECTED = 'JITSI_CLIENT_CONNECTED';
-export const JITSI_CLIENT_DISCONNECTED = 'JITSI_CLIENT_DISCONNECTED';
-export const JITSI_CLIENT_ERROR = 'JITSI_CLIENT_ERROR';
-
-export const RTC_ERROR = 'RTC_ERROR';
-
-export const JITSI_CONFERENCE_JOINED = 'JITSI_CONFERENCE_JOINED';
 
 export const REMOTE_TRACK_ADDED = 'REMOTE_TRACK_ADDED';
 export const REMOTE_TRACK_REMOVED = 'REMOTE_TRACK_ADDED';
@@ -16,16 +8,8 @@ export const LOCAL_STREAM_ENDED = 'LOCAL_STREAM_ENDED';
 
 export const CHANGE_CAMERA_SOURCE = 'CHANGE_CAMERA_SOURCE';
 export const CHANGE_MIC_SOURCE = 'CHANGE_MIC_SOURCE';
-export const CHANGE_CAMERA_FACING_MODE = 'CHANGE_CAMERA_FACING_MODE';
 
-export const TOGGLE_AUDIO_MUTED_STATE = 'TOGGLE_AUDIO_MUTED_STATE';
-export const TOGGLE_VIDEO_MUTED_STATE = 'TOGGLE_VIDEO_MUTED_STATE';
-
-export const PEER_JOINED = 'PEER_JOINED';
-export const PEER_LEFT = 'PEER_LEFT';
 export const PEER_CHANGED = 'PEER_CHANGED';
-export const DOMINANT_SPEAKER_CHANGED = 'DOMINANT_SPEAKER_CHANGED';
-export const MODERATOR_CHANGED = 'MODERATOR_CHANGED';
 
 export const UPDATE_PROFILE = 'UPDATE_PROFILE';
 
@@ -55,3 +39,4 @@ export const CAMERA_FACING_MODE = {
     ENVIRONMENT: 'environment',
     USER: 'user'
 };
+
