+++ features/reducers/index.android.js
@@ -2,21 +2,14 @@ require('../polyfills');
 
 import user from './user';
 import participants from './participants';
-import client from './client';
 import localTracks from './localTracks';
 import remoteTracks from './remoteTracks';
-import media from './media';
-
 
 const subReducers = {
-    client,
     user,
-    participants,
     localTracks,
-    remoteTracks,
-    media
+    remoteTracks
 };
-
 
 export default function reducer(state = {}, action) {
     let newState = {};
