+++ features/reducers/localTracks.js
@@ -8,9 +8,9 @@ import {
  */
 export default function (state = [], action) {
     switch (action.type) {
-        case LOCAL_TRACKS_CHANGED:
-            return action.tracks;
-        default:
-            return state;
+    case LOCAL_TRACKS_CHANGED:
+        return action.tracks;
+    default:
+        return state;
     }
 }
