+++ features/toolbar/actionTypes.js
@@ -1,0 +1,3 @@
+export const CHANGE_CAMERA_FACING_MODE = 'CHANGE_CAMERA_FACING_MODE';
+export const TOGGLE_AUDIO_MUTED_STATE = 'TOGGLE_AUDIO_MUTED_STATE';
+export const TOGGLE_VIDEO_MUTED_STATE = 'TOGGLE_VIDEO_MUTED_STATE';
