+++ features/welcome/actionTypes.js
@@ -1,0 +1,10 @@
+export const DOMINANT_SPEAKER_CHANGED = 'DOMINANT_SPEAKER_CHANGED';
+export const JITSI_CLIENT_CONNECTED = 'JITSI_CLIENT_CONNECTED';
+export const JITSI_CLIENT_CREATED = 'JITSI_CLIENT_CREATED';
+export const JITSI_CLIENT_DISCONNECTED = 'JITSI_CLIENT_DISCONNECTED';
+export const JITSI_CLIENT_ERROR = 'JITSI_CLIENT_ERROR';
+export const JITSI_CONFERENCE_JOINED = 'JITSI_CONFERENCE_JOINED';
+export const MODERATOR_CHANGED = 'MODERATOR_CHANGED';
+export const PEER_JOINED = 'PEER_JOINED';
+export const PEER_LEFT = 'PEER_LEFT';
+export const RTC_ERROR = 'RTC_ERROR';
